requirements.txt
apache-airflow-providers-google
google-cloud-bigquery
pandas
requests